Change to WebClient

+++ ScoreSaberPlaylistCreator/Classes/BeatsaverSong.cs
@@ -40,7 +40,7 @@ namespace ScoreSaberPlaylistCreator.Classes
         public string songSubName { get; set; }
         public string songAuthorName { get; set; }
         public string levelAuthorName { get; set; }
-        public int bpm { get; set; }
+        public double bpm { get; set; }
     }
 
     public class Stats
